chore: separate schemas from models
