+++ python/src/acp_sdk/models/__init__.py
@@ -1,2 +1,3 @@
 from acp_sdk.models.errors import *  # noqa: F403
 from acp_sdk.models.models import *  # noqa: F403
+from acp_sdk.models.schemas import *  # noqa: F403
