+++ python/src/acp_sdk/models/models.py
@@ -175,43 +175,7 @@ RunEvent = Union[
 ]
 
 
-class RunCreateRequest(BaseModel):
-    agent_name: AgentName
-    session_id: SessionId | None = None
-    input: Message
-    mode: RunMode = RunMode.SYNC
-
-
-class RunCreateResponse(Run):
-    pass
-
-
-class RunResumeRequest(BaseModel):
-    await_: AwaitResume = Field(alias="await")
-    mode: RunMode
-
-
-class RunResumeResponse(Run):
-    pass
-
-
-class RunReadResponse(Run):
-    pass
-
-
-class RunCancelResponse(Run):
-    pass
-
-
 class Agent(BaseModel):
     name: str
     description: str | None = None
     metadata: Metadata = Metadata()
-
-
-class AgentsListResponse(BaseModel):
-    agents: list[Agent]
-
-
-class AgentReadResponse(Agent):
-    pass
